@@ -5,12 +5,15 @@ namespace Mongo.Contexts;
 
 public interface IContext
 {
-    Task<RateForDate> GetRateForDate(CurrencyCodesEnum code, DateTime dateTime);
+    Task<RateForDate> GetRateForDate(CurrencyCodesEnum code, DateTime date);
+    
+    Task<DateTime> GetMostRecentDate();
+    
     Task<int> InsertRateForDate(RateForDate rateForDate);
-
+    
     Task<int> InsertRatesForDate(IEnumerable<RateForDate> ratesForDate);
+    
+    Task<int> DeleteRateById(string id);
 
-    Task<int> DeleteRateForDate(string id);
-
-    Task<int> RemoveAllRates(bool confirm = false);
+    Task<int> DeleteAllRates(bool confirm = false);
 }
