@@ -5,6 +5,6 @@ using RateGetters.Rates.Models.Enums;
 
 namespace CommandLayer.Queries.Currencies.Rates.GetForDateQuery
 {
-    public record GetForDateSpecification(CurrencyCodesEnum Code, DateTime DateTime);
+    public record GetForDateSpecification(CurrencyCodesEnum Code, DateTime Date);
     public record GetForDateQuery(GetForDateSpecification Specification) : IRequest<RateForDate>;
 }
